* commit '8d910e005dcbc24d5dd03aeb1bb01aa80f9bfccb':
  puppet: change the directory of certificate

+++ puppet.sh
@@ -8,15 +8,20 @@
 
 cert_name=$(cat /etc/hostname | tr -d '\n').pem
 rundir=/var/run/puppetlabs
-certdir=/etc/ssl/private/edgescale
+certdir=/data
 
-[ ! -h ${certdir}/certs/${cert_name} ] && [ -e ${certdir}/certs/edgescale.pem ] && \
-sudo ln -s ${certdir}/certs/edgescale.pem ${certdir}/certs/${cert_name}
+if [ ! -h ${certdir}/certs/${cert_name} ] && [ -e ${certdir}/certs/edgescale.pem ] ; then
+  sudo ln -s ${certdir}/certs/edgescale.pem ${certdir}/certs/${cert_name}
+fi
 
-[ ! -h ${certdir}/private_keys/${cert_name} ] && [ -e ${certdir}/private_keys/edgescale.key ] && \
-sudo ln -s ${certdir}/private_keys/edgescale.key ${certdir}/private_keys/${cert_name}
+if [ ! -h ${certdir}/private_keys/${cert_name} ] && [ -e ${certdir}/private_keys/edgescale.key ] ; then
+  sudo ln -s ${certdir}/private_keys/edgescale.key ${certdir}/private_keys/${cert_name}
+fi
 
 # kill puppet agent if it exists
-[ -e ${rundir}/agent.pid ] && sudo kill -s SIGINT $(<"${rundir}/agent.pid")
+if [ -e ${rundir}/agent.pid ] ; then
+  sudo kill -s SIGINT $(<"${rundir}/agent.pid")
+  sleep 1
+fi
 
 puppet agent --daemonize $@
